requirements.txt
spacy-huggingface_hub
spacy-lookups-data